+++ src/stable_hasher.rs
@@ -10,7 +10,7 @@ mod tests;
 
 /// Trait for retrieving the result of the stable hashing operation.
 pub trait StableHasherResult: Sized {
-    fn finish(hasher: StableHasher) -> Self;
+    fn finish(hash: (u64, u64)) -> Self;
 }
 
 /// When hashing something that ends up affecting properties like symbol names,
@@ -40,12 +40,7 @@ impl StableHasher {
 
     #[inline]
     pub fn finish<W: StableHasherResult>(self) -> W {
-        W::finish(self)
-    }
-
-    #[inline]
-    pub fn finalize(self) -> (u64, u64) {
-        self.state.finish128()
+        W::finish(self.state.finish128())
     }
 }